правки модуляции
